Moved viewmodel factories to another package.
