+++ app/src/main/java/hu/mczinke/nasa_apod_viewer/viewmodels/factories/MainViewModelFactory.kt
@@ -1,7 +1,9 @@
-package hu.mczinke.nasa_apod_viewer.viewmodels
+package hu.mczinke.nasa_apod_viewer.viewmodels.factories
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import hu.mczinke.nasa_apod_viewer.viewmodels.MainViewModel
+import hu.mczinke.nasa_apod_viewer.viewmodels.Repository
 
 class MainViewModelFactory(private val repository: Repository) : ViewModelProvider.Factory {
